Fixing Canvas bug, if height is zero canvas did not through exception. Moving StringParseTests to commands package. add CanvasSingleTests. this test caches the bug above.
